requirements.txt
# Project Genji Requirements
google-generativeai>=0.3.2
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
sqlalchemy>=2.0.0
prefect>=3.0.0
httpx>=0.25.0

# n8n Integration dependencies
flask>=2.3.0
requests>=2.31.0
